Fixed namespace on musit thing to conform to guidelines
