+++ service_musit_thing/conf/application.conf
@@ -37,7 +37,7 @@ slick.dbs.default {
     # Remember to configure connection pool for production
     # connectionPool = disabled
     driver = "org.h2.Driver"
-    url = "jdbc:h2:mem:play-service_example"
+    url = "jdbc:h2:mem:play-service-musit-thing"
   }
 
   numThreads = 20
